[WIP] bpf: Add options to install custom IP filter programs
Adds options IPIngressFilterBPF and IPEgressFilterBPF to install custom
BPF programs to a cgroup. The filters are (XXX: will be) loaded from
/sys/fs/bpf/systemd/<unit-name>/ip-{in,e}gress-filter.

The new options extend the bpf-firewall controller and looks to the new
CGroupContext variables ip_{in,e}gress_filter_bpf to see if the custom
programs or the simplified IPAccounting and IPAddress{Allow,Deny} need
to be installed.